Add files via upload
Step 1: Open Your Project
Open Power BI Desktop. Select the File menu and navigate to where the AdventureWorksSales.pbix file is saved. Select the file and click Open in the file explorer window to open the saved project in the Power BI Desktop application.

Step 2: Go to Data View
Use the icons on the vertical toolbar on the left side of the Power BI interface to switch to Data view. 

Tip: The Data view icon resembles a table. Data view allows you to see the data contained in your project. Select the Orders view on the right of the screen and take a moment to observe the first 10 records. Note which record has the highest Order Total value.

Step 3: Navigate to Model View
Use the icons on the vertical toolbar on the left side of the Power BI interface to switch to Model view. Tip: The Model view icon shows tables linked by connectors.

You should now see a diagram representing all the tables in there port, their fields, and how they are related.

Step 4: Select Relationships
The lines connecting the tables represent relationships. There is a relationship between the Customers and Orders tables. Observe the line connecting the Customers and Orders tables representing the many-to-many relationship you need to modify. 

Tip: The many-to-many relationship is depicted by asterisks on both sides of the line.

Double-click the line representing the relationship to access the Edit Relationship dialog.

In the Cardinality drop-down, change the selection from its current state to one-to-many.

Tip: One-to-many relationships are simpler and faster for Power BI to navigate when loading data and calculating results.

In the Cross filter direction drop-down, choose the Single option to reduce the complexity of the model and limit the direction in which filters are applied. 

Step 5: Review and save your Changes
In Model View, check the relationship connecting Customers to Orders. This should now display the number one (1) attached to the Customers table and an Asterix (*) attached to the Orders table. This visual confirmation ensures the relationships have been set to One-to-many as intended

Save your Power BI project to your local computer.

Tip: Make sure you select an appropriate project name and folder path.

Conclusion
When you optimize a data model, you are improving a report's performance and making the information more accessible. This facilitates efficient decision-making and contributes to the success of the organization.